Change the argument types to 64-bit unsigned integers

+++ include/inform/state_encoding.h
@@ -34,7 +34,7 @@ static uint64_t const inform_encoding_error = 0x8000000000000000;
  *
  * @see inform_decode
  */
-uint64_t inform_encode(uint64_t const *state, size_t n, int base);
+uint64_t inform_encode(uint64_t const *state, uint64_t n, uint64_t base);
 /**
  * Decode an integer into a sequence of values.
  *
@@ -54,4 +54,4 @@ uint64_t inform_encode(uint64_t const *state, size_t n, int base);
  *
  * @see inform_encode
  */
-uint64_t* inform_decode(uint64_t encoding, size_t n, int base);
+uint64_t* inform_decode(uint64_t encoding, uint64_t n, uint64_t base);
